ensure publican tmp junk is always cleaned before building
Even though publish.sh builds the whole site fresh in a temp dir, the
Publican build still happens in the beaker checkout which might have any
old junk left over in it. So we need to clean it out before doing the
build.

Ideally one day Publican will let us specify the source (and
destination) directories so that it can be built fresh in the temp dir
like everything else...