Add Fedora rawhide repos
The .treeinfo has version=rawhide, but the /etc/os-release file in a
rawhide install has version set to 20 (perhaps because it is already
branched). So, I guess let's keep both, since the server and client
repos will also work when rawhide hasn't been branched yet (around the
release of 20, but 21 not yet branched).

Change-Id: I18bdeab7728bcb54ff4cc8e4122f49682d6658e3