build-yum-repos.py: support pulling all packages from a tag
For the harness repos, which have their own dedicated tags in Beaker's
Koji, we don't need to maintain the package list here, Koji can be the
authoritative source. The script will just pull down all tagged builds
(with support for exclusions, so that we can keep rhts-devel out of the
harness repos).


Former-commit-id: a0cdd8e8cde857b6ee899334bc5114de3b4fa964