switch to Beaker Koji instead of sed-rhel-7 for RHEL7 server repos
This tag now contains the subset of EPEL7 packages which are needed to
satisfy Beaker's dependencies. This will become our "blessed" snapshot
of EPEL7 which we will test and deploy with. We can pull updates
explicitly from EPEL7 in future as needed.

The packages are also signed with our new GPG key.

Change-Id: I7df9e5911cd1811b6206f4d5cdf6d1d3e4c60143