keep packages in the same directory as the yum repo
Previously we had the little hack which would make the repodata refer to
the packages in ../../rpms/ as a way of only storing the packages on
disk once while still referencing them in multiple repos (for example,
beaker-common and beaker-client packages, which are included in both the
client and server repos).

However that assumes that the package filenames are always unique and
the packages are immutable -- but that is not really the case.
Particularly now that we are dealing with package signing, a given
beaker-client-23.2-1.el6.noarch.rpm could be either signed or unsigned
so not necessarily always identical.

This patches changes the repos to follow the traditional layout of
having all packages in the directory with the yum repos, the special
/yum/rpms/ directory will be removed. We can use other tools such as
hardlink(1) and duperemove(1) to save disk space for truly identical
bits.