improve fetching and writing of packages when building yum repos
Preserve the modtime of the remote file. This is just a nicety to
prevent unnecessary modtime changes if we build the repos on a fresh
Jenkins slave and then copy them to beaker-project.org.

Re-fetch any file whose size does not match what Koji says. This will
prevent us from accidentally using a partially downloaded file and
assuming that it's complete.

(We can't check hashes unfortunately because Koji only carries the RPM
payload hash, not a hash of the entire file, so it's not sensitive to
things like whether the package is signed or not.)

Move written files into place atomically with os.rename(). This avoids
leaving behind a truncated RPM with a legit name in case the download
fails or is interrupted partway through.


Former-commit-id: b6940a18e05303705bcacfd963d7c53c88e7af94