stop publishing server and client repos for EOL Fedora releases
Even though we aren't updating these repos, in general it is nicer to
keep them permanently around to avoid breaking old links and any old
clients who haven't upgraded yet.

But we are short on storage on beaker-project.org and this is one way we
can save some space for packages that are probably never going to be
needed again.

They will still be in our Koji instance for us to refer to and/or
resurrect later if needed.

Change-Id: Ic7e91db223c8739c21349352f21a9d49c6516000