catch the case where a symlink is left behind in Jenkins
We previously had Fedora26 symlinked to Fedora25. I removed the symlink
in git but it was still present in the Jenkins work dir (since we don't
clean that) so the script was building both repos into the same place
incorrectly. Better to just bail out in that case, so it can be fixed up
properly.