don't require every single package to match
Right now this check is failing for Fedora 26, because we have had
Fedora 26 harness repos for longer than Fedora 26 client repos,
therefore the harness repo has some older versions that are not present
in the client repo. Ultimately that situation is fine -- what we really
care about is that the latest version available, which yum is going to
pick for installation, matches across them both.

This actually takes it back to basically the same check we had before,
except we have to do the version comparison manually because
.returnNewestByName() doesn't let us separate out packages from the
client repo vs. packages from the harness repo to check against each
other.


Former-commit-id: 20b613da667f731dbf0b540e62894e95c3106149