switch back to Brew for RHEL6 server yum repo
We centralised all our builds and repos into our own team-specific Koji
instance several years ago, back when it was forbidden to have Fedora
and EPEL buildroots in Brew.

Now that moratorium is lifted we can switch back to Brew (primarily to
get access to real S/390 and POWER builders instead of very slow
emulated ones).

I am slowly working through the set of tags and importing/rebuilding as
we go... this is the first one to move back. The rest to come later.

Change-Id: I58e68f191f809db83ecfffd03e291a881e983d20