drop repos for EOL releases
These are moved to https://beaker-project.org/yum-archive/ with an HTTP
redirect from the old URLs.

This way, we don't need to permanently preserve the Koji tags for the
old releases, nor do we have to continually download and rsync the old
packages over. They will live unchanging on beaker-project.org.

Change-Id: I6b8dd36722a77537c5032fad2f2d74991092dbba